2002-05-11  Vladimir Makarov  <vmakarov@users.sourceforge.net>
	* DINO/d_heap.c (expand_vector): Fix bug with disp.

	* DINO/d_func.c (tree_heap_node_hash): Use size_t.

	* DINO/socket_int.c (h_errno): Check on macro.

+++ DINO/d_func.c
@@ -5219,7 +5219,7 @@ tree_heap_node_hash (hash_table_entry_t n)
 {
   struct tree_heap_node *node = ((struct tree_heap_node *) n);
 
-  return (unsigned) node->tree_node;
+  return (size_t) node->tree_node;
 }
 
 /* Equality of nodes. */
